requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.0.0
python-multipart==0.0.6